+++ kreogl/lights/VolumetricLightingParams.hpp
@@ -3,7 +3,7 @@
 namespace kreogl {
     struct VolumetricLightingParams {
         float scattering = .1f;
-        float nbSteps = 10.f;
+        int nbSteps = 10;
         float defaultStepLength = 25.f;
         float intensity = 2.5f;
     };
